Versión 1.21. Finalización de la lógica e interfaz de la página usuarios-administración. Reajuste de footer de todas las páginas y inclusión de la pestaña perfil en el menú de navegación admin.

+++ php/eliminarUser.php
@@ -3,36 +3,30 @@
 require_once '../php/conexionDB.php';
 // Conectar a la base de datos
 $conn = conectarDB();
-
 // Creamos la función para eliminar el usuario
 // Verificar si hay datos en $_POST
 if ($_SERVER["REQUEST_METHOD"] == "POST"){
-    
-    $idUserDelete = $_POST['idUserDelete'];
-    // Abre o crea un archivo de registro
-    $file = fopen('debug.log', 'a');
-
-    // Escribe el mensaje de debug
-    fwrite($file, "Usuario a borrar: " . $idUserDelete . PHP_EOL);
-
-    // Cierra el archivo
-    fclose($file);
-    $consulta = "DELETE FROM users_login WHERE idUser = $idUserDelete";
-    // Ejecutar la consulta
-    if (mysqli_query($conn, $consulta)) {
-            $response = array("success" => true, "message" => "El usuario ha sido eliminado");
-            // Abre o crea un archivo de registro
-            $file = fopen('debug.log', 'a');
-
-            // Escribe el mensaje de debug
-            fwrite($file, "Usuario eliminado: " . $idUserDelete . PHP_EOL);
-
-            // Cierra el archivo
-            fclose($file);
-            echo json_encode($response);
-        } else {
-            $response = array("success" => false, "message" => "Error al eliminar el usuario");
-            echo json_encode($response);
+    // Comprobación de que la variable isUserDelete tiene algún valor mayor que 0
+    if (isset($_POST['idUserDelete']) && $_POST['idUserDelete'] > 0){
+        $idUserDelete = $_POST['idUserDelete'];
+        $consulta = "DELETE FROM users_login WHERE idUser = $idUserDelete";
+        // Ejecutar la consulta
+        if (mysqli_query($conn, $consulta)) {
+            // Consulta para eliminar los datos personales del usuario
+            $consultaData = "DELETE FROM users_data WHERE idUser = $idUserDelete";
+            if(mysqli_query($conn, $consultaData)){
+                $response = array("success" => true, "message" => "El usuario ha sido eliminado");
+                echo json_encode($response);
+            }
+                
+            } else {
+                $response = array("success" => false, "message" => "Error al eliminar el usuario");
+                echo json_encode($response);
+        }
+    }else{
+        // Si no se recibe un usuario, indicamos al admin que seleccione un usuario
+        $response = array('success' => false, "message" => "Seleccione un usuario.");
+        echo json_encode($response);
     }
 }
 ?>
